refactor: organize global styles and clean up src root
