+++ src/App.tsx
@@ -21,7 +21,7 @@ import { ChordOverlayDock } from "./components/ChordOverlayDock/ChordOverlayDock
 import { VersionBadge } from "./components/VersionBadge/VersionBadge";
 import { MainLayoutWrapper } from "./components/MainLayoutWrapper/MainLayoutWrapper";
 import sharedStyles from "./components/shared.module.css";
-import "./App.css";
+import "./styles/App.css";
 
 const ExpandedControlsPanel = lazy(() =>
   import("./components/ExpandedControlsPanel/ExpandedControlsPanel").then((m) => ({
