+++ src/main.tsx
@@ -1,8 +1,8 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import './tokens.css'
-import './index.css'
-import './semantic.css'
+import './styles/tokens.css'
+import './styles/index.css'
+import './styles/semantic.css'
 import App from './App.tsx'
 import { ErrorBoundary, ErrorFallback } from './components/ErrorBoundary/ErrorBoundary'
